+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Mapbox welcomes participation and contributions from everyone.
 * Expose `LineLayer.lineCutoutOpacity` and `LineLayer.lineCutoutWidth` to make route lines visible through obstructing 3D buildings and other aboveground features.
 * Expose experimental `MapboxMap.setStyleLayerPropertiesAsync` and `MapboxMap.setStyleLayerPropertyAsync` (also available through `Style`) to allow setting properties from a any thread and asynchronously.
 
+## Bug fixes 🐞
+* Fix flyTo animation when zoom property is null or not changed
+
 # 11.15.0-rc.1 August 28, 2025
 
 ## Bug fixes 🐞